Refactored front-end for improved readability. Retested the options and improved commenting... Should generate docstrings soon
