+++ helpers/coord_extractor.py
@@ -1,0 +1,7 @@
+def coords(s):  
+    try:
+	sT = s.strip()
+        ra, dec = map(float, sT[1:len(sT)-1].split(','))
+        return ra, dec
+    except:
+        raise argparse.ArgumentTypeError("Coordinates must be ra,dec tupples")
